refactor: align naming and barrels with scanner
- ban non-submodule imports of effect/effect-app/@effect via oxlint
- drop addMissingImports vscode action
- rename *.controllers.ts to *.Controllers.ts; controllers.ts to routes.ts
- drop Response class wrapper in HelloWorld resource
- remove api services.ts/resources.ts barrels; use direct #services/<X> + #resources/<X>; namespace import * as XRsc
- drop frontend prelude ($$); inline submodule imports
- switch named effect/effect-app imports to namespace submodule imports across api/frontend/e2e/test
- deep submodule paths for @effect/platform-node, @effect/sql-sqlite-node, effect/unstable/http/FetchHttpClient
