--- api/src/controllers.ts
@@ -1,6 +1,0 @@
-// codegen:start {preset: barrel, include: ./*.controllers.ts, import: default}
-import accountsControllers from "./Accounts.controllers.js"
-import helloWorldControllers from "./HelloWorld.controllers.js"
-
-export { accountsControllers, helloWorldControllers }
-// codegen:end
