+++ api/src/lib/basicRuntime.ts
@@ -1,13 +1,20 @@
 import { reportError } from "@effect-app/infra/errorReporter"
 import { logJson } from "@effect-app/infra/logger/jsonLogger"
-import { NodeFileSystem } from "@effect/platform-node"
+import * as NodeFileSystem from "@effect/platform-node/NodeFileSystem"
 import * as Sentry from "@sentry/node"
 import dotenv from "dotenv"
-import { Cause, Console, Effect, Fiber, Layer, ManagedRuntime, pipe, References } from "effect-app"
+import * as Effect from "effect-app/Effect"
+import { pipe } from "effect-app/Function"
+import * as Layer from "effect-app/Layer"
+import * as Cause from "effect/Cause"
 import * as ConfigProvider from "effect/ConfigProvider"
+import * as Console from "effect/Console"
+import * as Fiber from "effect/Fiber"
 import * as Logger from "effect/Logger"
 import type * as LogLevel from "effect/LogLevel"
+import * as ManagedRuntime from "effect/ManagedRuntime"
 import type { ManagedRuntime as ManagedRuntimeType } from "effect/ManagedRuntime"
+import * as References from "effect/References"
 import * as Runtime from "effect/Runtime"
 import { DevTools } from "effect/unstable/devtools"
 import { TracingLive } from "./observability.js"
