+++ api/src/lib/layers.ts
@@ -1,11 +1,15 @@
 import { FakeSendgrid } from "@effect-app/infra/Emailer/fake"
 import { Sendgrid } from "@effect-app/infra/Emailer/Sendgrid"
 import { StoreMakerLayer } from "@effect-app/infra/Store/index"
-import { NodeServices } from "@effect/platform-node"
 import * as HttpClientNode from "@effect/platform-node/NodeHttpClient"
 import * as HttpNode from "@effect/platform-node/NodeHttpServer"
-import { SqliteClient } from "@effect/sql-sqlite-node"
-import { Context, Effect, Layer, Option, Redacted } from "effect-app"
+import * as NodeServices from "@effect/platform-node/NodeServices"
+import * as SqliteClient from "@effect/sql-sqlite-node/SqliteClient"
+import * as Context from "effect-app/Context"
+import * as Effect from "effect-app/Effect"
+import * as Layer from "effect-app/Layer"
+import * as Option from "effect-app/Option"
+import * as Redacted from "effect/Redacted"
 import fs from "fs"
 import { createServer } from "http"
