+++ api/src/lib/observability.ts
@@ -12,8 +12,12 @@ import { BatchSpanProcessor, ConsoleSpanExporter, NoopSpanProcessor } from "@ope
 import { SemanticResourceAttributes } from "@opentelemetry/semantic-conventions"
 import * as Sentry from "@sentry/node"
 import { SentryPropagator, SentrySampler, SentrySpanProcessor, setupEventContextTrace, wrapContextManagerClass } from "@sentry/opentelemetry"
-import { Config, Context, Effect, Layer, Redacted } from "effect-app"
+import * as Config from "effect-app/Config"
+import * as Context from "effect-app/Context"
+import * as Effect from "effect-app/Effect"
+import * as Layer from "effect-app/Layer"
 import { dropUndefinedT } from "effect-app/utils"
+import * as Redacted from "effect/Redacted"
 import fs from "fs"
 import tcpPortUsed from "tcp-port-used"
 import { baseConfig } from "../config.js"
