+++ api/src/main.ts
@@ -1,6 +1,9 @@
 import { faker } from "@faker-js/faker"
-import { type Config, Effect, Layer, Record } from "effect-app"
+import type * as Config from "effect-app/Config"
+import * as Effect from "effect-app/Effect"
 import { setFaker } from "effect-app/faker"
+import * as Layer from "effect-app/Layer"
+import * as Record from "effect/Record"
 import { api } from "./api.js"
 import { apiConfig } from "./config.js"
 import { runMain } from "./lib/basicRuntime.js"
