+++ api/src/resources/HelloWorld.ts
@@ -7,18 +7,20 @@ import { UserView } from "./views.js"
 const Req = TaggedRequestFor("HelloWorld")
 // codegen:end
 
-class Response extends S.Opaque<Response>()(S.Struct({
-  now: S.Date.withConstructorDefault,
-  echo: S.String,
-  state: S.String,
-  context: RequestContext,
-  currentUser: S.NullOr(UserView),
-  randomUser: UserView
-})) {}
-
 export class GetHelloWorld extends Req.Query<GetHelloWorld>()("GetHelloWorld", {
   echo: S.String
-}, { allowAnonymous: true, allowRoles: ["user"], success: Response }) {}
+}, {
+  allowAnonymous: true,
+  allowRoles: ["user"],
+  success: S.Struct({
+    now: S.Date.withConstructorDefault,
+    echo: S.String,
+    state: S.String,
+    context: RequestContext,
+    currentUser: S.NullOr(UserView),
+    randomUser: UserView
+  })
+}) {}
 
 export class SetState extends Req.Command<SetState>()("SetState", {
   state: S.String,
