+++ api/src/resources/lib/Userprofile.ts
@@ -1,6 +1,7 @@
 import { Role } from "#models/User"
-import { Context, S } from "effect-app"
+import * as Context from "effect-app/Context"
 import { UserProfileId } from "effect-app/ids"
+import * as S from "effect-app/Schema"
 
 // TODO: move back to services, and remove reference need in resources or frontend
 export class UserProfile extends Context.assignTag<UserProfile>("UserProfile")(
