+++ api/src/resources/lib/req.ts
@@ -1,6 +1,6 @@
-import { Layer } from "effect-app"
 import { makeRpcClient } from "effect-app/client"
 import { ApiClientFactory } from "effect-app/client/apiClientFactory"
+import * as Layer from "effect-app/Layer"
 import { AppMiddleware } from "./middleware.js"
 
 export const { TaggedRequestFor } = makeRpcClient(AppMiddleware)
