+++ api/src/routes.ts
@@ -1,0 +1,6 @@
+// codegen:start {preset: barrel, include: ./*.Controllers.ts, import: default}
+import accountsControllers from "./Accounts.Controllers.js"
+import helloWorldControllers from "./HelloWorld.Controllers.js"
+
+export { accountsControllers, helloWorldControllers }
+// codegen:end
