--- api/src/services.ts
@@ -1,6 +1,0 @@
-// codegen:start {preset: barrel, include: services/*.ts }
-export * from "./services/DBContext.js"
-export * from "./services/Events.js"
-export * from "./services/lib.js"
-export * from "./services/UserProfile.js"
-// codegen:end
