+++ api/src/services/DBContext/UserRepo.ts
@@ -1,13 +1,22 @@
 import { apiConfig } from "#config"
 import { RepoDefault } from "#lib/layers"
 import { User, UserFromIdResolver, type UserId } from "#models/User"
-import { Model } from "@effect-app/infra"
 import { NotFoundError, NotLoggedInError } from "@effect-app/infra/errors"
+import * as Model from "@effect-app/infra/Model"
 import { generate } from "@effect-app/infra/test"
-import { Array, Context, Effect, Exit, Layer, Option, pipe, Request, RequestResolver, S } from "effect-app"
+import * as Array from "effect-app/Array"
+import * as Context from "effect-app/Context"
+import * as Effect from "effect-app/Effect"
 import { fakerArb } from "effect-app/faker"
+import { pipe } from "effect-app/Function"
 import { UserProfileId } from "effect-app/ids"
+import * as Layer from "effect-app/Layer"
+import * as Option from "effect-app/Option"
+import * as S from "effect-app/Schema"
 import { Email } from "effect-app/Schema"
+import * as Exit from "effect/Exit"
+import * as Request from "effect/Request"
+import * as RequestResolver from "effect/RequestResolver"
 import fc from "fast-check"
 import { Q } from "../lib.js"
 import { UserProfile } from "../UserProfile.js"
