+++ api/test/some.test.ts
@@ -1,4 +1,4 @@
-import { HelloWorldRsc } from "#resources"
+import * as HelloWorldRsc from "#resources/HelloWorld"
 
 it("works", () => {
   console.log(HelloWorldRsc)
