+++ frontend/components/ErrorCause.vue
@@ -1,6 +1,8 @@
 <script setup lang="ts">
 import { useRuntimeConfig } from "#app"
-import { Cause, Match, Result } from "effect-app"
+import * as Cause from "effect/Cause"
+import * as Match from "effect/Match"
+import * as Result from "effect/Result"
 import type { SupportedErrors } from "effect-app/client/errors"
 
 defineProps<{ cause: Cause.Cause<unknown> }>()
