+++ frontend/components/FixedNuxtErrorBoundary.vue
@@ -14,7 +14,7 @@
 <script setup lang="ts">
 import { onNuxtReady, useNuxtApp, useRouter } from "#imports"
 import { captureException } from "@sentry/browser"
-import { Cause } from "effect-app"
+import * as Cause from "effect/Cause"
 import { CauseException } from "effect-app/client/errors"
 import { onErrorCaptured, shallowRef } from "vue"
