+++ frontend/composables/client.ts
@@ -8,7 +8,9 @@ import { I18n } from "@effect-app/vue/intl"
 import { makeClient } from "@effect-app/vue/makeClient"
 import * as Toast_ from "@effect-app/vue/toast"
 import { WithToast } from "@effect-app/vue/withToast"
-import { Effect, Layer, ManagedRuntime } from "effect-app"
+import * as Effect from "effect-app/Effect"
+import * as Layer from "effect-app/Layer"
+import * as ManagedRuntime from "effect/ManagedRuntime"
 import { useToast } from "vue-toastification"
 import type { RT } from "~/plugins/runtime"
 import { useIntl } from "./intl"
