+++ frontend/composables/useHelloWorld.ts
@@ -1,7 +1,8 @@
-import { HelloWorldRsc } from "#resources"
+import * as HelloWorldRsc from "#resources/HelloWorld"
+import { clientFor } from "./client"
 
 export const useHelloWorld = () => {
   const client = clientFor(HelloWorldRsc)
 
-  return {...client.helpers, client }
+  return { ...client.helpers, client }
 }
