+++ frontend/utils/observability.ts
@@ -5,7 +5,8 @@ import { BatchSpanProcessor } from "@opentelemetry/sdk-trace-web"
 import { browserTracingIntegration } from "@sentry/browser"
 import { SentryPropagator, SentrySpanProcessor } from "@sentry/opentelemetry-node"
 import * as Sentry from "@sentry/vue"
-import { Effect, Layer } from "effect-app"
+import * as Effect from "effect-app/Effect"
+import * as Layer from "effect-app/Layer"
 import { isErrorSilenced } from "effect-app/client/errors"
 import type { App } from "vue"
